AllMessages functionality implemented

+++ server/src/handlers/clientConnectionHandler.js
@@ -1,14 +1,16 @@
 exports.clientConnection = function (socket, wwebjsClient) {
     var loggedOut = false;
-    socket.on('log_out', (data) => {
+    socket.on('log_out', async (data) => {
         console.log("logout server")
-        wwebjsClient.destroy()
+        await wwebjsClient.destroy()
         socket.disconnect();
         loggedOut = true;
     })
     socket.on('disconnect', async () => {
         console.log('Socket Disconnected')
-        await wwebjsClient.destroy()
-        console.log('Wwebjs client destroyed!')
+        if (!loggedOut) {
+            await wwebjsClient.destroy()
+            console.log('Wwebjs client destroyed!')
+        }
     })
 }
